AI-Python/requirements.txt
fastapi==0.115.4
groq==0.11.0
openai==1.54.3
pydantic==2.9.2
python-dotenv==1.0.1
python-multipart==0.0.17
requests==2.32.3
uvicorn==0.32.0